requirements.txt
pandas
sqlalchemy
pyyaml
matplotlib
seaborn